+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2024-09-20
+
+### Added
+- Tooltip to display the variable type name on story editor dropdowns.
+
+### Changed
+- Nodes that generate their own children (for example Switch and Flow nodes with named ports) will auto align their children better than before.
+- Condition context menu will show "Inspect Script" instead of "Edit Script" when opened for a built-in condition.
+- Switch Node: Removed the warnings about no connected child during operation to avoid spamming the log.
+
+### Fixed
+- Text truncate issues on story editor dropdown elements.
+- Blackboard variables from added blackboard assets can now be dragged and dropped onto fields in nodes.
+- Fixed `CheckCollisionsInRadiusAction` failling to assign collided object if the BBV value was null.
+- Nodes with inherited named children (i.e. Node connections) weren't showing the base class' nodes.
+- Fixed Event Channel variables not appearing in link searches for non-event nodes.
+- Fixed the Self variable not working in Dynamic Subgraphs.
+
 ## [1.0.1] - 2024-09-18
 
 ### Fixed